% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_data}
\alias{cluster_data}
\title{Clusters a table of data}
\usage{
cluster_data(tab, col.names, k, method = c("clara", "flowsom"), ...)
}
\arguments{
\item{tab}{The input \code{data.frame}}

\item{col.names}{A vector specifying which columns of \code{tab} should be used for clustering}

\item{k}{The desired number of clusters}

\item{...}{Additional arguments to be passed to \code{cluster::clara}}
}
\value{
Returns \code{tab}, with an added column called \code{groups}, indicating cluster membership
}
\description{
This function clusters an input \code{data.frame} using the \code{cluster::clara} function
}
