% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_fcs_files}
\alias{cluster_fcs_files}
\title{Cluster FCS files}
\usage{
cluster_fcs_files(files.list, num.cores, col.names, num.clusters,
  asinh.cofactor, num.samples = 50, output.type = "legacy",
  output.dir = ".")
}
\arguments{
\item{files.list}{The files to cluster}

\item{num.cores}{Number of CPU cores to use}

\item{col.names}{A vector of column names indicating which columns should be used for clustering}

\item{num.clusters}{The desired number of clusters}

\item{asinh.cofactor}{Cofactor for asinh transformation (see \code{convert_fcs})}

\item{num.samples}{Number of samples to be used for the CLARA algorithm (see \code{cluster::clara})}

\item{output.dir}{Output directory}
}
\value{
Returns the list of files that have been clustered
}
\description{
Cluster FCS files
}
