% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_fcs_files_groups}
\alias{cluster_fcs_files_groups}
\title{Pool FCS files and cluster them}
\usage{
cluster_fcs_files_groups(
  files.list,
  num.cores,
  col.names,
  num.clusters,
  asinh.cofactor,
  num.samples = 50,
  downsample.to = 0,
  output.dir = ".",
  negative.values = "truncate",
  quantile.prob = 0.05,
  method = c("clara", "flowsom")
)
}
\arguments{
\item{files.list}{A named list of vectors detailing how the files should be pooled before clustering. Files in the same vector will
be pooled together. The name of the output is going to correspond to the name of the corresponding list element.}

\item{num.cores}{Number of CPU cores to use}

\item{col.names}{A vector of column names indicating which columns should be used for clustering}

\item{num.clusters}{The desired number of clusters}

\item{asinh.cofactor}{Cofactor for \code{asinh} transformation. If this is \code{NULL} no transformation is performed}

\item{num.samples}{Number of samples to be used for the CLARA algorithm (see \code{cluster::clara}).
Only used if \code{method = "clara"}}

\item{downsample.to}{The number of events that should be randomly sampled from each file before pooling. If this is 0, no sampling is performed}

\item{output.dir}{The name of the output directory, it will be created if it does not exist}

\item{negative.values}{How to deal with negative values in the data. If this is \code{NULL} negative values
are left as is. Otherwise two options are possible:
\itemize{
  \item{\code{truncate}}: Negative values will be truncated (i.e. replaced with 0)
  \item{\code{shift}}: The data will be shifted so that only \code{quantile.prob} of the values for each channel will
    be truncated to 0. This option is useful in cases where the range of data significantly extends
    in the negatives, for instance due to compensation.
}}

\item{quantile.prob}{Only used if \code{negative.value} is set to \code{shift}. The quantile of measurements
that are going to be truncated to 0. For instance if this is 0.05, the data will be shifted so that
only 5 percent of the values are negative and will be truncated to 0}
}
\value{
Returns either \code{NULL} or a \code{try-error} object if some error occurred during the computation
}
\description{
Pool FCS files and cluster them
}
