% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{convert_fcs}
\alias{convert_fcs}
\title{Convert a flowFrame to data.frame}
\usage{
convert_fcs(f, asinh.cofactor = NULL, clip.at.zero = T, compensate = T)
}
\arguments{
\item{f}{The \code{flowFrame} to convert}

\item{asinh.cofactor}{Cofactor for \code{asinh} transformation. If this is \code{NULL} no transformation is performed}

\item{clip.at.zero}{Wether to clip negative values (after transformation) at zero}

\item{compensate}{Wether to compensate the data using the compensation matrix embedded in the \code{flowFrame} (if any)}
}
\value{
Returns a \code{data.frame} corresponding to the data in \code{flowCore::exprs(f)} after compensation
  and transformation
}
\description{
This function converts a \code{flowFrame} object to an R data.frame, taking care of data
transformation and compensation.
}
\details{
The column names of the resulting data.frame will be derived from the channel descriptions (i.e. $PxS FCS keyword)
when possible, and the cannel names otherwise ($PxN)
}
