% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_common_columns}
\alias{get_common_columns}
\title{Get the columns that are common to a set of input tabular files}
\usage{
get_common_columns(files.list, file.type = c("txt", "fcs"))
}
\arguments{
\item{files.list}{A vector of input file names. If these are text files, each file should be a tab-separated table,
with the first row representing column headers}

\item{file.type}{The type of the files}
}
\value{
Returns a vector of column names that are present in all the files in \code{files.list}
}
\description{
Get the columns that are common to a set of input tabular files
}
