% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{get_stats_by_sample}
\alias{get_stats_by_sample}
\title{Calcularte sample-level cluster statistics}
\usage{
get_stats_by_sample(tab)
}
\arguments{
\item{tab}{A \code{data.frame}. Must contain two columns called \code{cellType} and \code{sample}, respectively indicating
the cluster membership and the sample of origin}
}
\value{
Returns a \code{data.frame} with the same columns as \code{tab}, plus sample-level statistics for every column in \code{tab}.
  If \code{tab} contains column \code{foo}, the returned value will contain columns \code{foo@sample1}, \code{foo@sample2} etc., containing the values
  of foo, as calculated only on the rows of \code{tab} that belong to \code{sample1}, \code{sample2} etc.
}
\description{
Calcularte sample-level cluster statistics
}
