% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{melt_cluster_results}
\alias{melt_cluster_results}
\title{Melt clustering results}
\usage{
melt_cluster_results(tab, features = NULL, transform.popsize = TRUE)
}
\arguments{
\item{tab}{The input data (i.e. the content of a pooled \code{clustered.txt} file)}

\item{features}{Optional. The features to extract. If not provided all features
will be extracted. Feature names in the original data are in the format
\code{feature@sample}}

\item{transform.popsize}{Whether to transform the popsize feature into proportion,
by normalizing for the total number of cells in a sample.}
}
\value{
Returns a \code{data.frame} of molten data
}
\description{
This function melts pooled clustering results using \code{reshape2::melt}
}
