% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{process_file}
\alias{process_file}
\title{Process an individual file for clustering}
\usage{
process_file(f, col.names, num.clusters, num.samples, asinh.cofactor,
  output.dir)
}
\arguments{
\item{f}{The file path}

\item{col.names}{A vector of column names indicating which columns should be used for clustering}

\item{num.clusters}{The desired number of clusters}

\item{num.samples}{Number of samples to be used for the CLARA algorithm (see \code{cluster::clara})}

\item{asinh.cofactor}{Cofactor for asinh transformation. If this is \code{NULL} no transformation is performed (see \code{convert_fcs})}

\item{output.dir}{The name of the output directory, it will be created if it does not exist}
}
\description{
Process an individual file for clustering
}
