% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{remove_empty_files}
\alias{remove_empty_files}
\title{Remove empty files}
\usage{
remove_empty_files(files.list, events.threshold = 0)
}
\arguments{
\item{files.list}{The list of FCS files to check, files with number of events \code{<= events.threshold}
will be removed}

\item{events.threshold}{A number. FCS files with number of events \code{<=} than this threshold will be removed}
}
\description{
This is an utility function to remove empty FCS files, which can cause problems with downstream analysis
}
