% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{shift_negative_values}
\alias{shift_negative_values}
\title{Shift negative values in a matrix}
\usage{
shift_negative_values(m, quantile.prob = 0.05)
}
\arguments{
\item{m}{The data matrix}

\item{quantile.prob}{The quantile probability to use}
}
\value{
Return the transformed data matrix
}
\description{
This function shifts negative values in a data matrix. For each column vector
the procedure is as follows:
\enumerate{
  \item A specific quantile is calculated from the vector
  \item If the quantile is negative then its absolute value is added to the vector
  \item Values that are still negative are truncated at 0
}
}
