% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{write_clustering_output}
\alias{write_clustering_output}
\title{Write clustering output}
\usage{
write_clustering_output(tab.medians, base.name, clustered.data, output.dir)
}
\arguments{
\item{tab.medians}{A \code{data.frame} containing median values for each cluster}

\item{base.name}{The base name for naming output files}

\item{clustered.data}{A \code{data.frame} containing the original data, with an exta column called \code{cellType} indicating
clustering membership}
}
\description{
Write clustering output
}
