% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_fcs_files}
\alias{cluster_fcs_files}
\title{Cluster FCS files}
\usage{
cluster_fcs_files(files.list, num.cores, col.names, num.clusters,
  asinh.cofactor, num.samples = 50, output.type = "file",
  output.dir = ".")
}
\arguments{
\item{files.list}{The files to cluster}

\item{num.cores}{Number of CPU cores to use}

\item{col.names}{A vector of column names indicating which columns should be used for clustering}

\item{num.clusters}{The desired number of clusters}

\item{asinh.cofactor}{Cofactor for asinh transformation. If this is \code{NULL} no transformation is performed (see \code{convert_fcs})}

\item{num.samples}{Number of samples to be used for the CLARA algorithm (see \code{cluster::clara})}

\item{output.type}{Either \code{"file"} or \code{"directory"}. See, Details}

\item{output.dir}{The directory in which all the output files (and directories) will be created}
}
\value{
Returns either \code{NULL} or a \code{try-error} object if some error occurred during the computation
}
\description{
Cluster individual FCS files, using multiple CPU cores if possible
}
\details{
This function can produce two types of output:
 \itemize{
   \item{\code{"file"}}: {Two files will be written, with names derived by appending \code{".clustered.txt"} and
     \code{".all_events.rds"} to the file names in \code{files.list}. The \code{".clustered.txt"} file is a tab-separated
     table of median values for each cluster. The \code{".all_events.rds"} file is an RDS file (readable with \code{base::readRDS})
     containing a \code{data.frame} with all the rows in the original input file and an additional column called
     \code{cellType}, indicating cluster membership.
   }
   \item{\code{"directory"}}: {In addition to the \code{".clustered.txt"} file described above, this mode will create a folder
     called \code{"clusters_data"}. This folder will contain a sub-folder for each input file, containing separate RDS files
     with the data in each cluster
   }
 }
}
