% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_fcs_files}
\alias{cluster_fcs_files}
\title{Cluster FCS files}
\usage{
cluster_fcs_files(
  files.list,
  num.cores,
  col.names,
  num.clusters,
  asinh.cofactor,
  num.samples = 50,
  output.dir = ".",
  negative.values = "truncate",
  quantile.prob = 0.05,
  method = c("clara", "flowsom")
)
}
\arguments{
\item{files.list}{The files to cluster}

\item{num.cores}{Number of CPU cores to use}

\item{col.names}{A vector of column names indicating which columns should be used for clustering}

\item{num.clusters}{The desired number of clusters}

\item{asinh.cofactor}{Cofactor for \code{asinh} transformation. If this is \code{NULL} no transformation is performed}

\item{num.samples}{Number of samples to be used for the CLARA algorithm (see \code{cluster::clara}).
Only used if \code{method = "clara"}}

\item{output.dir}{The name of the output directory, it will be created if it does not exist}

\item{negative.values}{How to deal with negative values in the data. If this is \code{NULL} negative values
are left as is. Otherwise two options are possible:
\itemize{
  \item{\code{truncate}}: Negative values will be truncated (i.e. replaced with 0)
  \item{\code{shift}}: The data will be shifted so that only \code{quantile.prob} of the values for each channel will
    be truncated to 0. This option is useful in cases where the range of data significantly extends
    in the negatives, for instance due to compensation.
}}

\item{quantile.prob}{Only used if \code{negative.value} is set to \code{shift}. The quantile of measurements
that are going to be truncated to 0. For instance if this is 0.05, the data will be shifted so that
only 5 percent of the values are negative and will be truncated to 0}

\item{The}{method to be used for clustering}
}
\value{
Returns either \code{NULL} or a \code{try-error} object if some error occurred during the computation
}
\description{
Cluster individual FCS files, using multiple CPU cores if possible
}
\details{
This function can produce two types of output:
 \itemize{
   \item{\code{"file"}}: {Two files will be written, with names derived by appending \code{".clustered.txt"} and
     \code{".all_events.rds"} to the file names in \code{files.list}. The \code{".clustered.txt"} file is a tab-separated
     table of median values for each cluster. The \code{".all_events.rds"} file is an RDS file (readable with \code{base::readRDS})
     containing a \code{data.frame} with all the rows in the original input file and an additional column called
     \code{cellType}, indicating cluster membership.
   }
   \item{\code{"directory"}}: {In addition to the \code{".clustered.txt"} file described above, this mode will create a folder
     called \code{"clusters_data"}. This folder will contain a sub-folder for each input file, containing separate RDS files
     with the data in each cluster
   }
 }
}
