% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_fcs_files_groups}
\alias{cluster_fcs_files_groups}
\title{Pool FCS files and cluster them}
\usage{
cluster_fcs_files_groups(files.list, num.cores, col.names, num.clusters,
  asinh.cofactor, num.samples = 50, downsample.to = 0, output.dir = ".")
}
\arguments{
\item{files.list}{A named list of vectors detailing how the files should be pooled before clustering. Files in the same vector will
be pooled together. The name of the output is going to correspond to the name of the corresponding list element.}

\item{num.cores}{Number of CPU cores to use}

\item{col.names}{A vector of column names indicating which columns should be used for clustering}

\item{num.clusters}{The desired number of clusters}

\item{asinh.cofactor}{Cofactor for asinh transformation. If this is \code{NULL} no transformation is performed (see \code{convert_fcs})}

\item{num.samples}{Number of samples to be used for the CLARA algorithm (see \code{cluster::clara})}

\item{downsample.to}{The number of events that should be randomly sampled from each file before pooling. If this is 0, no sampling is performed}

\item{output.dir}{The name of the output directory, it will be created if it does not exist}
}
\value{
Returns either \code{NULL} or a \code{try-error} object if some error occurred during the computation
}
\description{
Pool FCS files and cluster them
}
