% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_fcs_files_in_dir}
\alias{cluster_fcs_files_in_dir}
\title{Cluster all FCS files contained in a directory}
\usage{
cluster_fcs_files_in_dir(wd, ...)
}
\arguments{
\item{wd}{The directory containing the FCS files}

\item{...}{Arguments passed on to \code{cluster_fcs_files}
\describe{
  \item{num.cores}{Number of CPU cores to use}
  \item{col.names}{A vector of column names indicating which columns should be used for clustering}
  \item{num.clusters}{The desired number of clusters}
  \item{asinh.cofactor}{Cofactor for asinh transformation. If this is \code{NULL} no transformation is performed (see \code{convert_fcs})}
  \item{num.samples}{Number of samples to be used for the CLARA algorithm (see \code{cluster::clara})}
  \item{output.dir}{The name of the output directory, it will be created if it does not exist}
}}
}
\value{
Returns the list of files that have been clustered
}
\description{
Cluster all FCS files contained in a directory
}
