% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{convert_fcs}
\alias{convert_fcs}
\title{Convert a flowFrame to data.frame}
\usage{
convert_fcs(f, asinh.cofactor, transform.data = T, clip.at.zero = T,
  compensate = T)
}
\arguments{
\item{f}{The \code{flowFrame} to convert}

\item{asinh.cofactor}{Cofactor for \code{asinh} transformation}

\item{transform.data}{Whether to apply \code{asinh} transformation to the data}

\item{clip.at.zero}{Wether to clip negative values (after transformation) at zero}

\item{compensate}{Wether to compensate the data using the compensation matrix embedded in the \code{flowFrame} (if any)}
}
\description{
Convert a flowFrame to data.frame
}

