% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{process_file}
\alias{process_file}
\title{Process an individual file for clustering}
\usage{
process_file(f, col.names, num.clusters, num.samples, asinh.cofactor,
  output.dir, negative.values)
}
\arguments{
\item{f}{The file path}

\item{col.names}{A vector of column names indicating which columns should be used for clustering}

\item{num.clusters}{The desired number of clusters}

\item{num.samples}{Number of samples to be used for the CLARA algorithm (see \code{cluster::clara})}

\item{asinh.cofactor}{Cofactor for \code{asinh} transformation. If this is \code{NULL} no transformation is performed}

\item{output.dir}{The name of the output directory, it will be created if it does not exist}

\item{negative.values}{How to deal with negative values in the data. If this is \code{NULL} negative values
are left as is. Otherwise two options are possible:
\itemize{
  \item{\code{truncate}}: Negative values will be truncated (i.e. replaced with 0)
  \item{\code{shift}}: The data will be shifted so that only 5 percent of the values for each channel will
    be truncated to 0. This option is useful in cases where the range of data significantly extends
    in the negatives, for instance due to compensation.
}}
}
\description{
Process an individual file for clustering
}
