% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{process_files_groups}
\alias{process_files_groups}
\title{Process a group of files for clustering}
\usage{
process_files_groups(files, col.names, num.clusters, num.samples,
  asinh.cofactor, downsample.to, output.dir, negative.values,
  quantile.prob)
}
\arguments{
\item{files}{A vector of strings. The first string in the vector corresponds to the name to be used for the clustering output,
the remaining strings are the paths of the files that will be pooled together for clustering}

\item{col.names}{A vector of column names indicating which columns should be used for clustering}

\item{num.clusters}{The desired number of clusters}

\item{num.samples}{Number of samples to be used for the CLARA algorithm (see \code{cluster::clara})}

\item{asinh.cofactor}{Cofactor for \code{asinh} transformation. If this is \code{NULL} no transformation is performed}

\item{downsample.to}{The number of events that should be randomly sampled from each file before pooling. If this is 0, no sampling is performed}

\item{output.dir}{The name of the output directory, it will be created if it does not exist}

\item{negative.values}{How to deal with negative values in the data. If this is \code{NULL} negative values
are left as is. Otherwise two options are possible:
\itemize{
  \item{\code{truncate}}: Negative values will be truncated (i.e. replaced with 0)
  \item{\code{shift}}: The data will be shifted so that only \code{quantile.prob} of the values for each channel will
    be truncated to 0. This option is useful in cases where the range of data significantly extends
    in the negatives, for instance due to compensation.
}}

\item{quantile.prob}{Only used if \code{negative.value} is set to \code{shift}. The quantile of measurements
that are going to be truncated to 0. For instance if this is 0.05, the data will be shifted so that
only 5 percent of the values are negative and will be truncated to 0}
}
\description{
Process a group of files for clustering
}
