% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{process_files_groups}
\alias{process_files_groups}
\title{Process a group of files for clustering}
\usage{
process_files_groups(files, col.names, num.clusters, num.samples,
  asinh.cofactor, downsample.to, output.type, output.dir)
}
\arguments{
\item{files}{A vector of FCS file paths. Data from these files will be pooled and clustered together}

\item{col.names}{A vector of column names indicating which columns should be used for clustering}

\item{num.clusters}{The desired number of clusters}

\item{num.samples}{Number of samples to be used for the CLARA algorithm (see \code{cluster::clara})}

\item{asinh.cofactor}{Cofactor for asinh transformation (see \code{convert_fcs})}

\item{downsample.to}{The number of events that should be randomly sampled from each file before pooling. If this is 0, no sampling is performed}

\item{output.dir}{The name of the output directory}
}
\description{
Process a group of files for clustering
}
