% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{write_clustering_output}
\alias{write_clustering_output}
\title{Write clustering output}
\usage{
write_clustering_output(base.name, tab.medians, clustered.data,
  output.type = c("file", "directory"), output.dir)
}
\arguments{
\item{base.name}{The base name for naming output files}

\item{tab.medians}{A \code{data.frame} containing median values for each cluster}

\item{clustered.data}{A \code{data.frame} containing the original data, with an exta column called \code{cellType} indicating
clustering membership}

\item{output.type}{The type of output desired, see \code{cluster_fcs_files}}

\item{output.dir}{The output directory}
}
\description{
Write clustering output
}
